refs #5192 continued menu refactoring. not sure yet how good this idea is but seems to be already better than before
