+++ core/Plugin/Menu.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * Piwik - Open source web analytics
+ *
+ * @link http://piwik.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+namespace Piwik\Plugin;
+
+use Piwik\Menu\MenuAdmin;
+use Piwik\Menu\MenuMain;
+use Piwik\Menu\MenuTop;
+use Piwik\Menu\MenuUser;
+
+class Menu
+{
+    public function configureReportingMenu(MenuMain $menu)
+    {
+    }
+
+    public function configureTopMenu(MenuTop $menu)
+    {
+    }
+
+    public function configureAdminMenu(MenuAdmin $menu)
+    {
+    }
+
+    public function configureUserMenu(MenuUser $menu)
+    {
+    }
+
+}
