+++ plugins/CoreAdminHome/CoreAdminHome.php
@@ -12,12 +12,8 @@ use Piwik\DataAccess\ArchiveSelector;
 use Piwik\DataAccess\ArchiveTableCreator;
 use Piwik\Date;
 use Piwik\Db;
-use Piwik\Menu\MenuAbstract;
-use Piwik\Menu\MenuAdmin;
-use Piwik\Piwik;
 use Piwik\ScheduledTask;
 use Piwik\ScheduledTime;
-use Piwik\Settings\Manager as SettingsManager;
 use Piwik\Settings\UserSetting;
 
 /**
@@ -33,7 +29,6 @@ class CoreAdminHome extends \Piwik\Plugin
         return array(
             'AssetManager.getStylesheetFiles' => 'getStylesheetFiles',
             'AssetManager.getJavaScriptFiles' => 'getJsFiles',
-            'Menu.Admin.addItems'             => 'addMenu',
             'TaskScheduler.getScheduledTasks' => 'getScheduledTasks',
             'UsersManager.deleteUser'         => 'cleanupUser'
         );
@@ -85,27 +80,6 @@ class CoreAdminHome extends \Piwik\Plugin
         $jsFiles[] = "plugins/CoreAdminHome/javascripts/generalSettings.js";
         $jsFiles[] = "plugins/CoreHome/javascripts/donate.js";
         $jsFiles[] = "plugins/CoreAdminHome/javascripts/pluginSettings.js";
-    }
-
-    function addMenu(MenuAbstract $menu)
-    {
-        $menu->add('CoreAdminHome_MenuManage', null, "", Piwik::isUserHasSomeAdminAccess(), $order = 1);
-        $menu->add('CoreAdminHome_MenuDiagnostic', null, "", Piwik::isUserHasSomeAdminAccess(), $order = 10);
-        $menu->add('General_Settings', null, "", Piwik::isUserHasSomeAdminAccess(), $order = 5);
-        $menu->add('General_Settings', 'CoreAdminHome_MenuGeneralSettings',
-            array('module' => 'CoreAdminHome', 'action' => 'generalSettings'),
-            Piwik::isUserHasSomeAdminAccess(),
-            $order = 6);
-        $menu->add('CoreAdminHome_MenuManage', 'CoreAdminHome_TrackingCode',
-            array('module' => 'CoreAdminHome', 'action' => 'trackingCodeGenerator'),
-            Piwik::isUserHasSomeAdminAccess(),
-            $order = 4);
-
-        $menu->add('General_Settings', 'CoreAdminHome_PluginSettings',
-            array('module' => 'CoreAdminHome', 'action' => 'pluginSettings'),
-            SettingsManager::hasPluginsSettingsForCurrentUser(),
-            $order = 7);
-
     }
 
     function purgeOutdatedArchives()
