+++ plugins/CustomVariables/Menu.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Piwik - Open source web analytics
+ *
+ * @link http://piwik.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+namespace Piwik\Plugins\CustomVariables;
+
+use Piwik\Db;
+use Piwik\Menu\MenuReporting;
+
+/**
+ */
+class Menu extends \Piwik\Plugin\Menu
+{
+
+    public function confiugreReportingMenu(MenuReporting $menu)
+    {
+        $menu->add('General_Visitors', 'CustomVariables_CustomVariables', array('module' => 'CustomVariables', 'action' => 'index'), $display = true, $order = 50);
+    }
+
+}
