+++ plugins/Dashboard/Dashboard.php
@@ -32,8 +32,6 @@ class Dashboard extends \Piwik\Plugin
             'AssetManager.getJavaScriptFiles'        => 'getJsFiles',
             'AssetManager.getStylesheetFiles'        => 'getStylesheetFiles',
             'UsersManager.deleteUser'                => 'deleteDashboardLayout',
-            'Menu.Reporting.addItems'                => 'addMenus',
-            'Menu.Top.addItems'                      => 'addTopMenu',
             'Translate.getClientSideTranslationKeys' => 'getClientSideTranslationKeys'
         );
     }
@@ -195,37 +193,6 @@ class Dashboard extends \Piwik\Plugin
     public function encodeLayout($layout)
     {
         return Common::json_encode($layout);
-    }
-
-    public function addMenus(MenuAbstract $menu)
-    {
-        $menu->add('Dashboard_Dashboard', '', array('module' => 'Dashboard', 'action' => 'embeddedIndex', 'idDashboard' => 1), true, 5);
-
-        if (!Piwik::isUserIsAnonymous()) {
-            $login = Piwik::getCurrentUserLogin();
-
-            $dashboards = $this->getAllDashboards($login);
-
-            $pos = 0;
-            foreach ($dashboards as $dashboard) {
-                $menu->add('Dashboard_Dashboard', $dashboard['name'], array('module' => 'Dashboard', 'action' => 'embeddedIndex', 'idDashboard' => $dashboard['iddashboard']), true, $pos);
-                $pos++;
-            }
-        }
-    }
-
-    public function addTopMenu(MenuTop $menu)
-    {
-        $tooltip = false;
-        try {
-            $idSite = Common::getRequestVar('idSite');
-            $tooltip = Piwik::translate('Dashboard_TopLinkTooltip', Site::getNameFor($idSite));
-        } catch (Exception $ex) {
-            // if no idSite parameter, show no tooltip
-        }
-
-        $urlParams = array('module' => 'CoreHome', 'action' => 'index');
-        $menu->add('Dashboard_Dashboard', null, $urlParams, true, 1, $isHTML = false, $tooltip);
     }
 
     public function getJsFiles(&$jsFiles)
