+++ plugins/Events/Menu.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Piwik - Open source web analytics
+ *
+ * @link http://piwik.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+namespace Piwik\Plugins\Events;
+
+use Piwik\Menu\MenuReporting;
+
+/**
+ */
+class Menu extends \Piwik\Plugin\Menu
+{
+    public function configureReporingMenu(MenuReporting $menu)
+    {
+        $menu->add('General_Actions', 'Events_Events', array('module' => 'Events', 'action' => 'index'), true, 30);
+    }
+}
